requirements.txt
click
os
sys
mysql.connector